bundled_projects/project2/requirements.txt
streamlit>=1.55.0
openai>=2.28.0
PyPDF2>=3.0.1
python-dotenv>=1.2.2
